generalise functionality around roles and simplify
