+++ src/blockchain/network/participants/mod.rs
@@ -4,4 +4,4 @@ mod miners;
 mod shared;
 mod users;
 
-pub use self::{miners::miners_run, users::users_run};
+pub use self::{miners::miners_state_machine, shared::runner, users::users_state_machine};
